modified uri fetch - config execute pnpm

+++ src/public/javascripts/script.js
@@ -1,7 +1,8 @@
 console.log("works")
 
-//const uri = 'http://localhost:5000/api/v1/autores'
-const uri = 'https://deploy-blog-viajes.herokuapp.com/api/v1/autores'
+const uri = 'http://localhost:5000/api/v1/autores'
+//const uri = 'https://deploy-blog-viajes.herokuapp.com/api/v1/autores'
+
 let autores = []
 
 const getItems = () => {
@@ -42,6 +43,8 @@ const addItem = () => {
 }
 
 const deleteItem = (id) => {
+  const resultConfirm = confirm("¿Esta seguro de eliminar el autor seleccionado?")
+  if(!resultConfirm) return
   fetch(`${uri}/${id}`, {
     method: 'DELETE'
   })
